OneTwoTree

Documentation for OneTwoTree.

OneTwoTree.DecisionTree

OneTwoTree.Node

OneTwoTree.build_tree

OneTwoTree.collect_classes

OneTwoTree.equal

OneTwoTree.fit!

OneTwoTree.label_mean

OneTwoTree.label_mean

OneTwoTree.lessThanOrEqual

OneTwoTree.most_frequent_class

OneTwoTree.predict

OneTwoTree.printM

OneTwoTree.split_indices

OneTwoTree.split_indices

OneTwoTree.DecisionTree — Type

DecisionTree

A DecisionTree is a tree of Nodes. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.Node — Type

Node

A Node represents a decision in the Tree. It is a leaf with a prediction or has exactly one true and one false child and a decision function.

OneTwoTree.build_tree — Method

build_tree(features, labels, max_depth, ...)

Builds a decision tree from the given data using some algorithm (e.g. CART)

Arguments

tree::DecisionTree: the tree to be trained

dataset::Matrix{Union{Real, String}}: the training data

labels::Vector{Union{Real, String}}: the target labels

max_depth::Int: the maximum depth of the created tree

column_data::Bool: whether the datapoints are contained in dataset columnwise

OneTwoTree.collect_classes — Method

collect_classes(dataset, column)

Collect all unique classes among the specified column of the dataset.

Arguments

dataset::Matrix{Union{Real, String}}: the dataset to collect classes on

column::Int64: the index of the dataset column/feature to collect the classes on

OneTwoTree.equal — Function

equal

A basic categorical decision function for testing and playing around.

OneTwoTree.fit! — Method

fit!(tree, features, labels)

Train a decision tree on the given data using some algorithm (e.g. CART).

Arguments

tree::DecisionTree: the tree to be trained

dataset::Matrix{Union{Real, String}}: the training data

labels::Vector{Union{Real, String}}: the target labels

max_depth::Int: the maximum depth of the created tree

column_data::Bool: whether the datapoints are contained in dataset columnwise

OneTwoTree.label_mean — Method

label_mean(labels, indices)

Calculate the mean of a subset of numeric labels.

labels::Vector{Real}: the vector of numeric labels

indices::Vector{Int64}: the indices of the numeric labels to be considered

OneTwoTree.label_mean — Method

label_mean(labels, indices)

Calculate the mean of numeric labels.

Arguments

labels::Vector{Real}: the vector of numeric labels

OneTwoTree.lessThanOrEqual — Function

lessThanOrEqual

A basic numerical decision function for testing and playing around.

OneTwoTree.most_frequent_class — Method

most_frequent_class(labels, indices)

Determine the most frequent class among a subset of class labels.

Arguments

labels::Vector{String}: class labels

indices::Vector{Int64}: the indices of the class labels, to be considered/counted

OneTwoTree.predict — Method

predict

Traverses the tree for a given datapoint x and returns that trees prediction.

OneTwoTree.printM — Method

printM(M)

Print the matrix M:

Arguments

M::Matrix{Union{Real, String}}

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::Matrix{Union{Real, String}}: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing the dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint and returning a Bool

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn, decision_param, decision_feature)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::Matrix{Union{Real, String}}: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint, decisionparam and decisionfeature and returning a Bool

decision_param::Union{Real, String}: a parameter to the decision function. This can be a class name or a numeric threshold.

decision_feature::Int64: the index of the dimension of datapoints along which to split