OneTwoTree.DecisionTreeClassifier

OneTwoTree.DecisionTreeClassifier

OneTwoTree.DecisionTreeRegressor

OneTwoTree.DecisionTreeRegressor

OneTwoTree.ForestClassifier

OneTwoTree.ForestClassifier

OneTwoTree.ForestRegressor

OneTwoTree.ForestRegressor

OneTwoTree.calc_accuracy

OneTwoTree.fit!

OneTwoTree.fit!

OneTwoTree.gini_gain

OneTwoTree.information_gain

OneTwoTree.predict

OneTwoTree.predict

OneTwoTree.print_forest

OneTwoTree.print_tree

OneTwoTree.variance_gain

OneTwoTree.DecisionTreeClassifier — Type

DecisionTreeClassifier <: AbstractDecisionTree

A DecisionTreeClassifier is a tree of decision nodes. It can predict classes based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data.

OneTwoTree.DecisionTreeClassifier — Method

DecisionTreeClassifier(; root=nothing, max_depth=-1)

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Type

DecisionTreeRegressor

A DecisionTreeRegressor is a tree of decision nodes. It can predict function values based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data.

OneTwoTree.DecisionTreeRegressor — Method

DecisionTreeRegressor(; root=nothing, max_depth=-1)

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.ForestClassifier — Type

ForestClassifier <: AbstractForest

Forest for classification problems

Fields:

trees::Vector{DecisionTreeClassifier}: A vector of decision trees.

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

OneTwoTree.ForestClassifier — Method

ForestClassifier(;n_trees::Int, n_features_per_tree::Int, max_depth::Int)

Constructs a ForestClassifier instance.

Arguments:

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

Returns:

A ForestClassifier instance.

OneTwoTree.ForestRegressor — Type

ForestRegressor <: AbstractForest

Forest for regression problems

Fields:

trees::Vector{DecisionTreeRegressor}: A vector of regression trees.

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

OneTwoTree.ForestRegressor — Method

ForestRegressor(;n_trees::Int, n_features_per_tree::Int, max_depth::Int)

Constructs a ForestRegressor instance.

Arguments:

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

Returns:

A ForestRegressor instance.

OneTwoTree.calc_accuracy — Method

calc_accuracy(labels::AbstractArray, predictions::AbstractArray)

Calculates the accuracy of the predictions compared to the labels.

OneTwoTree.fit! — Method

fit!(tree::AbstractDecisionTree, features::AbstractMatrix{S}, labels::AbstractVector{T}; splitting_criterion=nothing, column_data=false) where {S, T<:Union{Real, String}}

Train a decision tree on the given data using some algorithm (e.g. CART).

Arguments

tree::AbstractDecisionTree: the tree to be trained

dataset::AbstractMatrix{S}: the training data

labels::AbstractVector{T}: the target labels

splitting_criterion::Function: a function indicating some notion of gain from splitting a node. If not provided, default criteria for classification and regression are used.

column_data::Bool: whether the datapoints are contained in dataset columnwise

(OneTwoTree provides the following splitting criteria for classification: ginigain, informationgain; and for regression: variance_gain. If you'd like to define a splitting criterion yourself, you need to consider the following:

The function must calculate a 'gain'-value for a split of a node, meaning that larger values are considered better.

The function signature must conform to my_func(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector),

where parent_labels is a set of datapoint labels, which is split into two subsets true_child_labels & false_child_labels by some discriminating function. (Each label in parent_labels is contained in exactly one of the two subsets.)

OneTwoTree.fit! — Method

fit!(forest::AbstractForest, dataset::AbstractMatrix{S}, labels::AbstractVector{T}; splitting_criterion=nothing, column_data=false) where {S, T<:Union{Real, String}}

Trains each tree in the forest on randomly drawn subsets of test features and corresponding test labels.

Arguments

forest::AbstractForest: the forest to be trained

dataset::AbstractMatrix{S}: the training data

labels::AbstractVector{T}: the target labels

splitting_criterion: a function indicating some notion of gain from splitting a node. If not provided, default criteria for classification and regression are used.

column_data::Bool: whether the datapoints are contained in dataset columnwise

(OneTwoTree provides the following splitting criteria for classification: ginigain, informationgain; and for regression: variance_gain. If you'd like to define a splitting criterion yourself, you need to consider the following:

The function must calculate a 'gain'-value for a split of a node, meaning that larger values are considered better.

The function signature must conform to my_func(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector),

where parent_labels is a set of datapoint labels, which is split into two subsets true_child_labels & false_child_labels by some discriminating function. (Each label in parent_labels is contained in exactly one of the two subsets.)

OneTwoTree.gini_gain — Method

gini_gain(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector)::Float64

This function calculates the gain in Gini impurity for a split in a decision tree. The split is characterized by the partition of the parentlabels into truechildlabels and falsechild_labels according to some discriminant function.

Arguments:

parent_labels: A vector of data labels (e.g., classes or numerical values in the case of regression).

true_child_labels: A vector of a subset of data labels contained in parent_labels.

false_child_labels: A vector of a subset of data labels contained in parent_labels.

Returns:

The gain in Gini impurity of the split.

OneTwoTree.information_gain — Method

information_gain(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector)::Float64

This function calculates the information gain for a split in a decision tree. The split is characterized by the partition of the parentlabels into truechildlabels and falsechild_labels according to some discriminant function.

Arguments:

parent_labels: A vector of data labels (e.g., classes or numerical values in the case of regression).

true_child_labels: A vector of a subset of data labels contained in parent_labels.

false_child_labels: A vector of a subset of data labels contained in parent_labels.

Returns:

The information gain of the split.

OneTwoTree.predict — Method

predict(tree::AbstractDecisionTree, X::Union{AbstractMatrix, AbstractVector})

Traverses the tree for given datapoints and returns that trees prediction.

Arguments

tree::AbstractDecisionTree: the tree to predict with

X::Union{AbstractMatrix, AbstractVector}: the data to predict on

OneTwoTree.predict — Method

predict(forest::AbstractForest, X::Union{AbstractMatrix{S}, AbstractVector{S}}) where S<:Union{Real, String}

Outputs the forest-prediction for a given datapoint X. The prediction is based on the aggregation of the tree decisions. For aggregation in a regression scenario the mean is used. For aggregation in a classification scenario the most voted class label is used.

Arguments:

forest::AbstractForest: The trained forest.

X::Union{AbstractMatrix{S}, AbstractVector{S}}: The input data for which a prediction is searched.

Returns:

Predictions for the input data X, aggregated across all trees in the forest.

Errors:

Raises an error if the forest contains no trained trees.

OneTwoTree.print_forest — Method

print_forest(forest::AbstractForest; io::IO=stdout)

Prints the numerated trees of a forest.

Arguments:

forest::AbstractForest: The forest to be printed.

io::IO=stdout: (Optional) The I/O stream for printing

OneTwoTree.print_tree — Method

print_tree(tree::AbstractDecisionTree; io::IO=stdout)

Returns a textual visualization of the decision tree.

Arguments

tree::AbstractDecisionTree The DecisionTree instance to print.

io::IO=stdout: (Optional) The I/O stream for printing

Example output:

x < 28.0 ?
├─ False: x == 161.0 ?
│  ├─ False: 842
│  └─ True: 2493
└─ True: 683

OneTwoTree.variance_gain — Method

variance_gain(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector)::Float64

This function calculates the variance gain for a split in a decision tree. The split is characterized by the partition of the parentlabels into truechildlabels and falsechild_labels according to some discriminant function.

Arguments:

parent_labels: A vector of data labels (e.g., classes or numerical values in the case of regression).

true_child_labels: A vector of a subset of data labels contained in parent_labels.

false_child_labels: A vector of a subset of data labels contained in parent_labels.

Returns:

The variance gain of the split.