OneTwoTree

Documentation for OneTwoTree.

OneTwoTree.DecisionTree

OneTwoTree.DecisionTree

OneTwoTree.Node

OneTwoTree.build_tree

OneTwoTree.fit!

OneTwoTree.lessThan

OneTwoTree.tree_prediction

OneTwoTree.DecisionTree — Type

DecisionTree

A DecisionTree is a tree of Nodes. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTree — Method

Initialises a decision tree model.

Arguments

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.Node — Type

Node

A Node represents a decision in the Tree. It is a leaf with a prediction or has exactly one true and one false child and a decision function.

OneTwoTree.build_tree — Method

build_tree(features, labels, max_depth, ...)

Builds a decision tree from the given data using some algorithm (e.g. CART)

Arguments

tree::DecisionTree: the tree to be trained

X::Array{Float64,2}: the training data

y::Array{Float64,1}: the target labels

OneTwoTree.fit! — Method

fit!(tree, features, labels)

Train a decision tree on the given data using some algorithm (e.g. CART).

Arguments

tree::DecisionTree: the tree to be trained

X::Array{Float64,2}: the training data

y::Array{Float64,1}: the target labels

OneTwoTree.lessThan — Function

lessThan

A basic decision function for testing and playing around.

OneTwoTree.tree_prediction — Method

tree_prediction

Traverses the tree for a given datapoint x and returns that trees prediction.