OneTwoTree

Documentation for OneTwoTree.

OneTwoTree.DecisionTreeClassifier

OneTwoTree.DecisionTreeClassifier

OneTwoTree.DecisionTreeRegressor

OneTwoTree.DecisionTreeRegressor

OneTwoTree.Node

OneTwoTree.build_tree

OneTwoTree.calc_accuracy

OneTwoTree.calc_depth

OneTwoTree.collect_classes

OneTwoTree.equal

OneTwoTree.fit!

OneTwoTree.is_leaf

OneTwoTree.label_mean

OneTwoTree.label_mean

OneTwoTree.lessThanOrEqual

OneTwoTree.load_data

OneTwoTree.most_frequent_class

OneTwoTree.predict

OneTwoTree.printM

OneTwoTree.print_tree

OneTwoTree.split_indices

OneTwoTree.split_indices

OneTwoTree.DecisionTreeClassifier — Type

DecisionTreeClassifier

A DecisionTreeClassifier is a tree of decision nodes. It can predict classes based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeClassifier — Method

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Type

DecisionTreeRegressor

A DecisionTreeRegressor is a tree of decision nodes. It can predict function values based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Method

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.Node — Type

Node

A Node represents a decision in the Tree. It is a leaf with a prediction or has exactly one true and one false child and a decision function.

OneTwoTree.build_tree — Method

build_tree(features, labels, max_depth, ...)

Builds a decision tree from the given data using some algorithm (e.g. CART)

Arguments

tree::AbstractDecisionTree: the tree to be trained

dataset::Matrix{Union{Real, String}}: the training data

labels::Vector{Union{Real, String}}: the target labels

max_depth::Int: the maximum depth of the created tree

column_data::Bool: whether the datapoints are contained in dataset columnwise

OneTwoTree.calc_accuracy — Method

calc_accuracy(labels, predictions)

Calculates the accuracy of the predictions compared to the labels.

OneTwoTree.calc_depth — Method

depth(tree)

Traverses the tree and returns the maximum depth.

OneTwoTree.collect_classes — Method

collect_classes(dataset, column)

Collect all unique classes among the specified column of the dataset.

Arguments

dataset::Matrix{Union{Real, String}}: the dataset to collect classes on

column::Int64: the index of the dataset column/feature to collect the classes on

OneTwoTree.equal — Method

equal

A basic categorical decision function for testing and playing around.

OneTwoTree.fit! — Method

fit!(tree, features, labels)

Train a decision tree on the given data using some algorithm (e.g. CART).

Arguments

tree::AbstractDecisionTree: the tree to be trained

dataset::Matrix{Union{Real, String}}: the training data

labels::Vector{Union{Real, String}}: the target labels

column_data::Bool: whether the datapoints are contained in dataset columnwise

OneTwoTree.is_leaf — Method

is_leaf(node)

Do you seriously expect a description for this?

OneTwoTree.label_mean — Method

label_mean(labels, indices)

Calculate the mean of a subset of numeric labels.

labels::Vector{Real}: the vector of numeric labels

indices::Vector{Int64}: the indices of the numeric labels to be considered

OneTwoTree.label_mean — Method

label_mean(labels, indices)

Calculate the mean of numeric labels.

Arguments

labels::Vector{Real}: the vector of numeric labels

OneTwoTree.lessThanOrEqual — Method

lessThanOrEqual

A basic numerical decision function for testing and playing around.

OneTwoTree.load_data — Method

load_data(name::String)

Load a preconfigured dataset from a CSV file.

Arguments

name::String: the name of the dataset to load

Example

load_data("fashion_mnist_1000")

OneTwoTree.most_frequent_class — Method

most_frequent_class(labels, indices)

Determine the most frequent class among a subset of class labels.

Arguments

labels::Vector{Union{Int, String}}: class labels

indices::Vector{Int64}: the indices of the class labels, to be considered/counted

OneTwoTree.predict — Method

predict

Traverses the tree for a given datapoint x and returns that trees prediction.

Arguments

tree::AbstractDecisionTree: the tree to predict with

X::Union{Matrix{S}, Vector{S}: the data to predict on

OneTwoTree.printM — Method

printM(M)

Print the matrix M:

Arguments

M::Matrix{Union{Real, String}}

OneTwoTree.print_tree — Method

print_tree(tree::AbstractDecisionTree)

Prints a textual visualization of the decision tree. For each decision node, it displays the condition, and for each leaf, it displays the prediction.

Arguments

tree::AbstractDecisionTree The DecisionTree instance to print.

Example output:

x < 28 ? ├─ False: y < 161 ? ├─ False: 842 └─ True: 2493 └─ True: 683

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::Matrix{Union{Real, String}}: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing the dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint and returning a Bool

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn, decision_param, decision_feature)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::Matrix{Union{Real, String}}: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint, decisionparam and decisionfeature and returning a Bool

decision_param::Union{Real, String}: a parameter to the decision function. This can be a class name or a numeric threshold.

decision_feature::Int64: the index of the dimension of datapoints along which to split