OneTwoTree.AbstractForest — Type

Base type for classification and regression forests.

Forests are collections of trees which aggregate their decisions. We use this abstract type to differentiate tree construction intricacies in the fit! function

OneTwoTree.Decision — Type

Decision

A structure representing a decision with a function and its parameter.

Parameters

fn::Function: The decision function.

param::Union{Number, String}: The parameter for the decision function.

Number: for comparison functions (e.g. x < 5.0)

String: for True/False functions (e.g. x == "red" or x != 681)

feature::Int64: The index of the feature to compare.

OneTwoTree.DecisionTreeClassifier — Type

DecisionTreeClassifier

A DecisionTreeClassifier is a tree of decision nodes. It can predict classes based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeClassifier — Method

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Type

DecisionTreeRegressor

A DecisionTreeRegressor is a tree of decision nodes. It can predict function values based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Method

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.ForestClassifier — Type

Forest for classification problems

Fields:

trees::Vector{DecisionTreeClassifier}: A vector of decision trees.

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

OneTwoTree.ForestClassifier — Method

ForestClassifier(;n_trees::Int, n_features_per_tree::Int, max_depth::Int)

Constructs a ForestClassifier instance.

Arguments:

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

Returns:

A ForestClassifier instance.

OneTwoTree.ForestRegressor — Type

Forest for regression problems

Fields:

trees::Vector{DecisionTreeRegressor}: A vector of regression trees.

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

OneTwoTree.ForestRegressor — Method

ForestRegressor(;n_trees::Int, n_features_per_tree::Int, max_depth::Int)

Constructs a ForestRegressor instance.

Arguments:

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

Returns:

A ForestRegressor instance.

OneTwoTree.Node — Type

Node

A Node represents a decision in the Tree. It is a leaf with a prediction or has exactly one true and one false child and a decision function.

Base.show — Method

Displays a string representation of the forest when shown in the REPL or other I/O streams.

Arguments:

io::IO: The I/O stream.

forest::AbstractForest: The forest to be displayed.

OneTwoTree._decision_to_string — Method

_decision_to_string(d::DecisionFn)

Returns a string representation of the decision function.

Arguments

d::DecisionFn: The decision function to convert to a string.

OneTwoTree._forest_to_string — Method

_forest_to_string(forest::AbstractForest)

Converts the forest to a string.

Arguments:

forest::AbstractForest: The forest to be represented.

Returns:

A string with numerated text representations of the trees in the forest

OneTwoTree._node_to_string — Method

_node_to_string(node::Node, prefix::String, is_true_child::Bool, indentation::String)

Recursive helper function to stringify the decision tree structure.

Arguments

node: The current node to print.

is_true_child: Boolean indicating if the node is a true branch child.

indentation: The current indentation.

OneTwoTree._tree_to_string — Function

tree_to_string(tree::AbstractDecisionTree)

Returns a textual visualization of the decision tree.

Arguments

tree::AbstractDecisionTree The DecisionTree instance to print.

Example output:

x < 28.0 ? ├─ False: x == 161.0 ? │ ├─ False: 842 │ └─ True: 2493 └─ True: 683

OneTwoTree._verify_fit!_args — Method

Some guards to ensure the input data is valid for training a tree.

OneTwoTree.calc_accuracy — Method

calc_accuracy(labels, predictions)

Calculates the accuracy of the predictions compared to the labels.

OneTwoTree.calc_depth — Method

depth(tree)

Traverses the tree and returns the maximum depth.

OneTwoTree.collect_classes — Method

collect_classes(dataset, column)

Collect all unique classes among a subset of the specified column of the dataset.

Arguments

dataset::AbstractMatrix: the dataset to collect classes on

indices::Vector{Int64}: the indices of the numeric labels to be considered

column::Int64: the index of the dataset column/feature to collect the classes on

OneTwoTree.entropy — Method

entropy(features::AbstractVector) -> Float64

This function calculates the entropy of set of lables

Arguments:

labels: A vector of labels

Returns:

The entropy H(X) = -sum^n{i=1} P(xi) log2(P(xi)) with X beeing the labels and n the number of elements in X and P() beeing the Probability

OneTwoTree.equal — Method

equal

A basic categorical decision function for testing and playing around.

OneTwoTree.fit! — Method

fit!(tree::AbstractDecisionTree, features::AbstractMatrix, labels::Vector{T}; splitting_criterion=nothing, column_data=false) where {T<:Union{Number, String}}

Train a decision tree on the given data using some algorithm (e.g. CART).

Arguments

tree::AbstractDecisionTree: the tree to be trained

dataset::AbstractMatrix: the training data

labels::Vector{Union{Number, String}}: the target labels

splitting_criterion: a function indicating some notion of gain from splitting a node. If not provided, default criteria for classification and regression are used.

column_data::Bool: whether the datapoints are contained in dataset columnwise

(OneTwoTree provides the following splitting criteria for classification: ginigain, informationgain; and for regression: variance_gain. If you'd like to define a splitting criterion yourself, you need to consider the following:

The function must calculate a 'gain'-value for a split of a node, meaning that larger values are considered better.

The function signature must conform to my_func(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector) where parentlabels is a set of datapoint labels, which is split into two subsets truechildlabels & falsechildlabels by some discriminating function. (Each label in parentlabels is contained in exactly one of the two subsets.)

OneTwoTree.fit! — Method

fit!(forest::AbstractForest, dataset::AbstractMatrix, labels::Vector{T}; splitting_criterion=nothing, column_data=false) where {T<:Union{Number, String}}

Trains each tree in the forest on randomly drawn subsets of test features and corresponding test labels.

Arguments

forest::AbstractForest: the forest to be trained

dataset::AbstractMatrix: the training data

labels::Vector{Union{Number, String}}: the target labels

splitting_criterion: a function indicating some notion of gain from splitting a node. If not provided, default criteria for classification and regression are used.

column_data::Bool: whether the datapoints are contained in dataset columnwise

(OneTwoTree provides the following splitting criteria for classification: ginigain, informationgain; and for regression: variance_gain. If you'd like to define a splitting criterion yourself, you need to consider the following:

The function must calculate a 'gain'-value for a split of a node, meaning that larger values are considered better.

The function signature must conform to my_func(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels::AbstractVector) where parentlabels is a set of datapoint labels, which is split into two subsets truechildlabels & falsechildlabels by some discriminating function. (Each label in parentlabels is contained in exactly one of the two subsets.)

OneTwoTree.get_random_features — Method

get_random_features(features::Matrix{S}, labels::Vector{T}, n_features::Int)

Returns random features and their corresponding labels from the given dataset.

Arguments:

features::Matrix{S}: The feature matrix.

labels::Vector{T}: The labels vector.

n_features::Int: The number of features to draw randomly.

Returns:

A tuple (random_features, random_labels) containing the randomly drawn features and labels.

OneTwoTree.gini_gain — Method

gini_gain(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels) -> Float64

This function calculates the gain in Gini impurity for a split in a decision tree. The split is characterized by the partition of the parentlabels into truechildlabels and falsechild_labels according to some discriminant function.

Arguments:

parent_labels: A vector of data labels (e.g., classes or numerical values in the case of regression).

true_child_labels: A vector of a subset of data labels contained in parent_labels.

false_child_labels: A vector of a subset of data labels contained in parent_labels.

Returns:

The gain in Gini impurity of the split.

OneTwoTree.information_gain — Method

information_gain(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels) -> Float64

This function calculates the information gain for a split in a decision tree. The split is characterized by the partition of the parentlabels into truechildlabels and falsechild_labels according to some discriminant function.

Arguments:

parent_labels: A vector of data labels (e.g., classes or numerical values in the case of regression).

true_child_labels: A vector of a subset of data labels contained in parent_labels.

false_child_labels: A vector of a subset of data labels contained in parent_labels.

Returns:

The information gain of the split.

OneTwoTree.is_leaf — Method

is_leaf(node)

Do you seriously expect a description for this?

OneTwoTree.label_mean — Method

label_mean(labels, indices)

Calculate the mean of a subset of numeric labels.

labels::Vector{Real}: the vector of numeric labels

indices::Vector{Int64}: the indices of the numeric labels to be considered

OneTwoTree.less_than_or_equal — Method

less_than_or_equal

A basic numerical decision function for testing and playing around.

OneTwoTree.most_frequent_class — Method

most_frequent_class(labels, indices)

Determine the most frequent class among a subset of class labels.

Arguments

labels::Vector{Union{Int, String}}: class labels

indices::Vector{Int64}: the indices of the class labels, to be considered/counted

OneTwoTree.predict — Method

predict

Traverses the tree for a given datapoint x and returns that trees prediction.

Arguments

tree::AbstractDecisionTree: the tree to predict with

X::Union{AbstractMatrix, AbstractVector: the data to predict on

OneTwoTree.predict — Method

predict(forest::AbstractForest, X::Union{Matrix{S}, Vector{S}}) where S<:Union{Real, String}

Outputs the forest-prediction for a given datapoint X. The prediction is based on the aggregation of the tree decisions. For agregation in a regression scenario the mean is used. For agregation in a classification scenario the most voted class label is used.

Arguments:

forest::AbstractForest: The trained forest.

X::Union{Matrix{S}, Vector{S}}: The input data for which a prediction is searched.

Returns:

Predictions for the input data X, aggregated across all trees in the forest.

Errors:

Raises an error if the forest contains no trained trees.

OneTwoTree.print_forest — Method

print_forest(forest::AbstractForest)

Prints the numerated trees of a forest.

Arguments:

forest::AbstractForest: The forest to be printed.

io::IO=stdout: (Optional) The I/O stream for printing

OneTwoTree.print_tree — Method

print_tree(tree::AbstractDecisionTree)

Returns a textual visualization of the decision tree.

Arguments

tree::AbstractDecisionTree The DecisionTree instance to print.

Example output:

x < 28.0 ? ├─ False: x == 161.0 ? │ ├─ False: 842 │ └─ True: 2493 └─ True: 683

OneTwoTree.should_split — Method

should_split(N, splitting_gain, max_depth)

Determines whether to split the node N given.

Arguments

N::Node: Node that may be split. N contains further fields relevant to the decision like the best splitting decision functtion and maximum tree depth.

splitting_gain::Float64: The gain in the splitting criterion metric of N after it's optimal split.

max_depth::Int64: The maximum depth of the tree N is part of.

OneTwoTree.split — Method

split(N)

Determine the optimal split of a node. Handles numerical and categorical data and labels.

Arguments

N::Node: The node to be split. All additional information for the split calculation (e.g. dataset, labels, node_data) is contained in N.

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn, decision_param, decision_feature)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::AbstractMatrix: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint, decisionparam and decisionfeature and returning a Bool

decision_param::Union{Real, String}: a parameter to the decision function. This can be a class name or a numeric threshold.

decision_feature::Int64: the index of the dimension of datapoints along which to split

OneTwoTree.variance_gain — Method

variance_gain(parent_labels::AbstractVector, true_child_labels::AbstractVector, false_child_labels) -> Float64

This function calculates the variance gain for a split in a decision tree. The split is characterized by the partition of the parentlabels into truechildlabels and falsechild_labels according to some discriminant function.

Arguments:

parent_labels: A vector of data labels (e.g., classes or numerical values in the case of regression).

true_child_labels: A vector of a subset of data labels contained in parent_labels.

false_child_labels: A vector of a subset of data labels contained in parent_labels.

Returns:

The variance gain of the split.

OneTwoTree.verify_forest_args — Method

verify_forest_args(n_trees::Int, n_features_per_tree::Int, max_depth::Int)

Verifies the validity of arguments used to initialize a forest.

Arguments:

n_trees::Int: The number of trees in the forest.

n_features_per_tree::Int: The number of randomly drawn features used for each tree.

max_depth::Int: The upper bound for the depth of each tree.

Errors:

Raises an error if any of the arguments are less than or equal to zero.