OneTwoTree.Decision — Type

Decision

A structure representing a decision with a function and its parameter.

Parameters

fn::Function: The decision function.

param::Union{Number, String}: The parameter for the decision function.

Number: for comparison functions (e.g. x < 5.0)

String: for True/False functions (e.g. x == "red" or x != 681)

feature::Int64: The index of the feature to compare.

OneTwoTree.DecisionTreeClassifier — Type

DecisionTreeClassifier

A DecisionTreeClassifier is a tree of decision nodes. It can predict classes based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeClassifier — Method

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Type

DecisionTreeRegressor

A DecisionTreeRegressor is a tree of decision nodes. It can predict function values based on the input data. In addition to a root node it holds meta informations such as max_depth etc. Use fit(tree, features, labels) to create a tree from data

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.DecisionTreeRegressor — Method

Initialises a decision tree model.

Arguments

root::Union{Node, Nothing}: the root node of the decision tree; nothing if the tree is empty

max_depth::Int: maximum depth of the decision tree; no limit if equal to -1

OneTwoTree.Node — Type

Node

A Node represents a decision in the Tree. It is a leaf with a prediction or has exactly one true and one false child and a decision function.

OneTwoTree._decision_to_string — Method

_decision_to_string(d::DecisionFn)

Returns a string representation of the decision function.

Arguments

d::DecisionFn: The decision function to convert to a string.

OneTwoTree._node_to_string — Method

_node_to_string(node::Node, prefix::String, is_true_child::Bool, indentation::String)

Recursive helper function to stringify the decision tree structure.

Arguments

node: The current node to print.

is_true_child: Boolean indicating if the node is a true branch child.

indentation: The current indentation.

OneTwoTree._tree_to_string — Function

tree_to_string(tree::AbstractDecisionTree)

Returns a textual visualization of the decision tree.

Arguments

tree::AbstractDecisionTree The DecisionTree instance to print.

Example output:

x < 28.0 ? ├─ False: x == 161.0 ? │ ├─ False: 842 │ └─ True: 2493 └─ True: 683

OneTwoTree._verify_fit!_args — Method

Some guards to ensure the input data is valid for training a tree.

OneTwoTree.calc_accuracy — Method

calc_accuracy(labels, predictions)

Calculates the accuracy of the predictions compared to the labels.

OneTwoTree.calc_depth — Method

depth(tree)

Traverses the tree and returns the maximum depth.

OneTwoTree.collect_classes — Method

collect_classes(dataset, column)

Collect all unique classes among the specified column of the dataset.

Arguments

dataset::AbstractMatrix: the dataset to collect classes on

column::Int64: the index of the dataset column/feature to collect the classes on

OneTwoTree.collect_classes — Method

collect_classes(dataset, column)

Collect all unique classes among a subset of the specified column of the dataset.

Arguments

dataset::AbstractMatrix: the dataset to collect classes on

indices::Vector{Int64}: the indices of the numeric labels to be considered

column::Int64: the index of the dataset column/feature to collect the classes on

OneTwoTree.equal — Method

equal

A basic categorical decision function for testing and playing around.

OneTwoTree.fit! — Method

fit!(tree, features, labels)

Train a decision tree on the given data using some algorithm (e.g. CART).

Arguments

tree::AbstractDecisionTree: the tree to be trained

dataset::AbstractMatrix: the training data

labels::Vector{Union{Real, String}}: the target labels

`splitting_criterion: a function indicating some notion of gain from splitting a node.

column_data::Bool: whether the datapoints are contained in dataset columnwise

(OneTwoTree provides the following splitting criteria for classification: ginigain, informationgain; and for regression: variance_gain. If you'd like to define a splitting criterion yourself, you need to consider the following:

The function must calculate a 'gain'-value for a split of a node, meaning that larger values are considered better.

The function signature must conform to my_func(features::AbstractMatrix, labels::AbstractVector, node_data::Vector{Int64}, decision_fn::Function, decision_param::Union{Real, String}, decision_feature::Int64)) where features is the total data matrix, labels the total label vector and node_data is an index array, that indexes into the former to define a subset of the data to be considered. The last three values define a split of the data.

You can get the split labelsets for your computation by taking analogous steps as in the OneTwoTree.split_indices method.)

OneTwoTree.information_gain — Method

information_gain(features::AbstractVector) -> Float64

This function calculates the information gain spliting criterion

Arguments:

parent_labels: A vector of all considerd labels to be split

child_1_labels: A vector of labels of the first spliting part

child_2_labels: A vector of labels of the other spliting part

Returns:

The information gain for given split as Float64.

calculated as follows: Information gain = entropy(parent) - [weightes] * entropy(children)

OneTwoTree.information_gain — Method

information_gain(features::AbstractVector) -> Float64

This function calculates the information gain spliting criterion

Arguments:

parent_labels: A vector of all considerd labels to be split

child_1_labels: A vector of labels of the first spliting part

child_2_labels: A vector of labels of the other spliting part

Returns:

The information gain for given split as Float64.

calculated as follows: Information gain = entropy(parent) - [weightes] * entropy(children)

OneTwoTree.is_leaf — Method

is_leaf(node)

Do you seriously expect a description for this?

OneTwoTree.label_mean — Method

label_mean(labels, indices)

Calculate the mean of a subset of numeric labels.

labels::Vector{Real}: the vector of numeric labels

indices::Vector{Int64}: the indices of the numeric labels to be considered

OneTwoTree.less_than_or_equal — Method

less_than_or_equal

A basic numerical decision function for testing and playing around.

OneTwoTree.most_frequent_class — Method

most_frequent_class(labels, indices)

Determine the most frequent class among a subset of class labels.

Arguments

labels::Vector{Union{Int, String}}: class labels

indices::Vector{Int64}: the indices of the class labels, to be considered/counted

OneTwoTree.predict — Method

predict

Traverses the tree for a given datapoint x and returns that trees prediction.

Arguments

tree::AbstractDecisionTree: the tree to predict with

X::Union{AbstractMatrix, AbstractVector: the data to predict on

OneTwoTree.printM — Method

printM(M)

Print the matrix M:

Arguments

M::Matrix{Union{Real, String}}

OneTwoTree.print_tree — Method

print_tree(tree::AbstractDecisionTree)

Returns a textual visualization of the decision tree.

Arguments

tree::AbstractDecisionTree The DecisionTree instance to print.

Example output:

x < 28.0 ? ├─ False: x == 161.0 ? │ ├─ False: 842 │ └─ True: 2493 └─ True: 683

OneTwoTree.should_split — Method

should_split(N, splitting_gain, max_depth)

Determines whether to split the node N given.

Arguments

N::Node: Node that may be split. N contains further fields relevant to the decision like the best splitting decision functtion and maximum tree depth.

splitting_gain::Float64: The gain in the splitting criterion metric of N after it's optimal split.

max_depth::Int64: The maximum depth of the tree N is part of.

OneTwoTree.split — Method

split(N)

Determine the optimal split of a node. Handles numerical and categorical data and labels.

Arguments

N::Node: The node to be split. All additional information for the split calculation (e.g. dataset, labels, node_data) is contained in N.

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::AbstractMatrix: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing the dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint and returning a Bool

OneTwoTree.split_indices — Method

split_indices(dataset, node_data, decision_fn, decision_param, decision_feature)

Split the dataset indices contained in node_data into two sets via the decision function.

Arguments

dataset::AbstractMatrix: the dataset to split on (the datapoints are assumed to be contained rowwise)

node_data::Vector{Int64}: the index list, indexing dataset, to be split

decision_fn::Function: the decision function taking as input a datapoint, decisionparam and decisionfeature and returning a Bool

decision_param::Union{Real, String}: a parameter to the decision function. This can be a class name or a numeric threshold.

decision_feature::Int64: the index of the dimension of datapoints along which to split